Refactor code for consistency: Update variable declarations for clarity and remove unnecessary whitespace, switch to explicit typing, add version checker.

+++ Config/Settings.cs
@@ -9,10 +9,7 @@ internal static class Settings
 
     internal static void IniFileSetup()
     {
-        if (!File.Exists("plugins/LSPDFR/GRIDWATCH.ini"))
-        {
-            File.Create("plugins/LSPDFR/GRIDWATCH.ini").Close();
-        }
+        if (!File.Exists("plugins/LSPDFR/GRIDWATCH.ini")) File.Create("plugins/LSPDFR/GRIDWATCH.ini").Close();
 
         IniReflector.Read(UserConfig, true);
         ValidateIniValues();
@@ -31,33 +28,33 @@ internal static class Settings
 
 internal class Config
 {
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: 85)]
-    public int ReadChance;
+    [IniReflectorValue("General_Settings", defaultValue: false)]
+    public bool DebugModeEnabled;
 
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: 1500)]
-    public int ScanInterval;
+    [IniReflectorValue("General_Settings", defaultValue: true)]
+    public bool EnableBlips;
 
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: 60000)]
-    public int ShotspotterPollRate;
-
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: 5)]
-    public int ShotspotterChance;
-
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: 20)]
-    public int ShotspotterFalseAlarmChance;
-
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: 5)]
+    [IniReflectorValue("General_Settings", defaultValue: 5)]
     public int MaxCamerasPerScan;
 
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: true)]
-    public bool EnableBlips;
-    
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: Keys.O)]
+    [IniReflectorValue("General_Settings", defaultValue: Keys.O)]
     public Keys MenuKey;
 
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: Keys.LControlKey)]
+    [IniReflectorValue("General_Settings", defaultValue: Keys.LControlKey)]
     public Keys MenuModifierKey;
 
-    [IniReflectorValue(sectionName: "General_Settings", defaultValue: false)]
-    public bool DebugModeEnabled;
+    [IniReflectorValue("General_Settings", defaultValue: 85)]
+    public int ReadChance;
+
+    [IniReflectorValue("General_Settings", defaultValue: 1500)]
+    public int ScanInterval;
+
+    [IniReflectorValue("General_Settings", defaultValue: 5)]
+    public int ShotspotterChance;
+
+    [IniReflectorValue("General_Settings", defaultValue: 20)]
+    public int ShotspotterFalseAlarmChance;
+
+    [IniReflectorValue("General_Settings", defaultValue: 60000)]
+    public int ShotspotterPollRate;
 }
